// src/app/models/demandestate.ts
export class DemandeState {

    idDemandeState: number;
    stateDemande: string;
    date: Date;
    isActive: boolean;
    comment: string;
    fkDemande: number;
    fkUser: number;
    isAccepted: boolean;
    isModified: boolean;
    // fkDemande: string;
    //

    codeSubsidary: string;
    // comment: string;
    // date: Date;
    dateDemande: Date;
    demandeId: number;
    designation: string;
    fkCategorie: number;

    fkMode: number;
    fkSubsidiary: number;
    //  fkUser: number;
    fullName: string;
    //  idDemandeState: number;

    // isActive: boolean;
    labelSubsidary: string;
    libelleRole: string;
    lieu: string;
    login: string;
    matricule: string;
    partie1: string;
    partie2: string;
    password: string;
    poste: string;
    //  stateDemande: string;
    typeCat: string;
    typeDemande: string;
    labelDemande: string;
    montant: number;

}